Move Redux code into store folder and create a store creation function
