+++ client/src/index.tsx
@@ -1,22 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from "react-redux";
-import { createStore, applyMiddleware } from "redux";
-import { composeWithDevTools } from '@redux-devtools/extension';
-import reducers from "./reducers";
-import thunk from "redux-thunk";
-
 import reportWebVitals from './reportWebVitals';
- 
+
+
+import createPreconfiguredStore from './store/createPresetStore';
 import App from "./components/App";
  
-const store = createStore(reducers, composeWithDevTools(applyMiddleware(thunk)));
-
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
 root.render(
-  <Provider store={store}>
+  <Provider store={createPreconfiguredStore()}>
     <App />
   </Provider>
 );
