+++ client/src/store/createPresetStore.ts
@@ -1,0 +1,13 @@
+import { createStore, applyMiddleware } from "redux";
+import { composeWithDevTools } from '@redux-devtools/extension';
+import thunk from "redux-thunk";
+
+import reducers from "./reducers";
+
+const createPreconfiguredStore = (initialState?: any) => createStore(
+  reducers,
+  initialState,
+  composeWithDevTools(applyMiddleware(thunk))
+);
+
+export default createPreconfiguredStore;
