+++ client/src/store/reducers/index.ts
@@ -3,8 +3,10 @@ import { reducer as formReducer} from "redux-form";
 import userReducer from "./userReducer";
 import authReducer from "./authReducer";
 
-export default combineReducers({
+const rootReducer = combineReducers ({
   users: userReducer,
   form: formReducer,
-  auth: authReducer,
+  auth: authReducer
 });
+
+export default rootReducer;
